requirements.txt
opencv-python
numpy
scikit-image
Pillow
tqdm
